requirements.txt
tabulate==0.9.0
pytest-cov==6.2.1